refactor: moved existing webview tests to legacy folder
