+++ src/test/suite/views/webview-app/legacy/components/app.test.tsx
@@ -8,15 +8,15 @@ import { createStore } from 'redux';
 import {
   initialState,
   rootReducer,
-} from '../../../../../views/webview-app/legacy/store/store';
+} from '../../../../../../views/webview-app/legacy/store/store';
 import App, {
   App as NotConnectedApp,
-} from '../../../../../views/webview-app/legacy/components/app';
-import OverviewPage from '../../../../../views/webview-app/legacy/components/overview-page/overview-page';
+} from '../../../../../../views/webview-app/legacy/components/app';
+import OverviewPage from '../../../../../../views/webview-app/legacy/components/overview-page/overview-page';
 import {
   CONNECTION_STATUS,
   MESSAGE_TYPES,
-} from '../../../../../views/webview-app/extension-app-message-constants';
+} from '../../../../../../views/webview-app/extension-app-message-constants';
 
 describe('App Component Test Suite', () => {
   describe('when rendered', () => {
