+++ src/test/suite/views/webview-app/legacy/components/atlas-cta/atlas-cta.test.tsx
@@ -8,9 +8,9 @@ import { Provider } from 'react-redux';
 import {
   initialState,
   rootReducer,
-} from '../../../../../../views/webview-app/legacy/store/store';
-import AtlasCTA from '../../../../../../views/webview-app/legacy/components/atlas-cta/atlas-cta';
-import { VSCODE_EXTENSION_SEGMENT_ANONYMOUS_ID } from '../../../../../../views/webview-app/extension-app-message-constants';
+} from '../../../../../../../views/webview-app/legacy/store/store';
+import AtlasCTA from '../../../../../../../views/webview-app/legacy/components/atlas-cta/atlas-cta';
+import { VSCODE_EXTENSION_SEGMENT_ANONYMOUS_ID } from '../../../../../../../views/webview-app/extension-app-message-constants';
 
 describe('Resources Panel Component Test Suite', () => {
   describe('when rendered', () => {
