+++ src/test/suite/views/webview-app/legacy/components/connect-form/connection-form.test.tsx
@@ -2,13 +2,13 @@ import assert from 'assert';
 import * as React from 'react';
 import { shallow } from 'enzyme';
 
-import { ConnectionForm } from '../../../../../../views/webview-app/legacy/components/connect-form/connection-form';
-import FormActions from '../../../../../../views/webview-app/legacy/components/form/form-actions';
-import { CONNECTION_FORM_TABS } from '../../../../../../views/webview-app/legacy/store/constants';
-import GeneralTab from '../../../../../../views/webview-app/legacy/components/connect-form/general-tab/general-tab';
-import SSLTab from '../../../../../../views/webview-app/legacy/components/connect-form/ssl-tab/ssl-tab';
-import SSHTunnelTab from '../../../../../../views/webview-app/legacy/components/connect-form/ssh-tab/ssh-tunnel-tab';
-import AdvancedTab from '../../../../../../views/webview-app/legacy/components/connect-form/advanced-tab/advanced-tab';
+import { ConnectionForm } from '../../../../../../../views/webview-app/legacy/components/connect-form/connection-form';
+import FormActions from '../../../../../../../views/webview-app/legacy/components/form/form-actions';
+import { CONNECTION_FORM_TABS } from '../../../../../../../views/webview-app/legacy/store/constants';
+import GeneralTab from '../../../../../../../views/webview-app/legacy/components/connect-form/general-tab/general-tab';
+import SSLTab from '../../../../../../../views/webview-app/legacy/components/connect-form/ssl-tab/ssl-tab';
+import SSHTunnelTab from '../../../../../../../views/webview-app/legacy/components/connect-form/ssh-tab/ssh-tunnel-tab';
+import AdvancedTab from '../../../../../../../views/webview-app/legacy/components/connect-form/advanced-tab/advanced-tab';
 
 describe('Webview Connection Form Component Test Suite', () => {
   test('it shows a connection form', () => {
