+++ src/test/suite/views/webview-app/legacy/components/connect-form/general-tab/authentication/authentication.test.tsx
@@ -2,11 +2,11 @@ import assert from 'assert';
 import * as React from 'react';
 import { shallow } from 'enzyme';
 
-import { Authentication } from '../../../../../../../../views/webview-app/legacy/components/connect-form/general-tab/authentication/authentication';
-import RadioBoxGroup from '../../../../../../../../views/webview-app/legacy/components/form/radio-box-group/radio-box-group';
-import MongodbAuthentication from '../../../../../../../../views/webview-app/legacy/components/connect-form/general-tab/authentication/mongodb-authentication';
-import AUTH_STRATEGIES from '../../../../../../../../views/webview-app/legacy/connection-model/constants/auth-strategies';
-import ScramSha256 from '../../../../../../../../views/webview-app/legacy/components/connect-form/general-tab/authentication/scram-sha-256';
+import { Authentication } from '../../../../../../../../../views/webview-app/legacy/components/connect-form/general-tab/authentication/authentication';
+import RadioBoxGroup from '../../../../../../../../../views/webview-app/legacy/components/form/radio-box-group/radio-box-group';
+import MongodbAuthentication from '../../../../../../../../../views/webview-app/legacy/components/connect-form/general-tab/authentication/mongodb-authentication';
+import AUTH_STRATEGIES from '../../../../../../../../../views/webview-app/legacy/connection-model/constants/auth-strategies';
+import ScramSha256 from '../../../../../../../../../views/webview-app/legacy/components/connect-form/general-tab/authentication/scram-sha-256';
 
 describe('Authentication Component Test Suite', () => {
   test('it shows a select for the authentication method', () => {
