+++ src/test/suite/views/webview-app/legacy/components/connect-helper/connect-helper.test.tsx
@@ -5,12 +5,12 @@ import { mount } from 'enzyme';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 
-import ConnectHelper from '../../../../../../views/webview-app/legacy/components/connect-helper/connect-helper';
-import type { AppState } from '../../../../../../views/webview-app/legacy/store/store';
+import ConnectHelper from '../../../../../../../views/webview-app/legacy/components/connect-helper/connect-helper';
+import type { AppState } from '../../../../../../../views/webview-app/legacy/store/store';
 import {
   initialState,
   rootReducer,
-} from '../../../../../../views/webview-app/legacy/store/store';
+} from '../../../../../../../views/webview-app/legacy/store/store';
 
 describe('Connect Helper Component Test Suite', () => {
   describe('when rendered', () => {
